requirements.txt
docling
ollama
pdf2doi
pyyaml
tqdm
streamlit
streamlit-pdf-viewer

